+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Made the `levenshtein`, `levenshteinNorm`, `damerauLevenshtein`, and
   `demerauLevenshtein` more efficient.
 
+* Added `jaro` and `jaroWinkler` functions.
+
 ## Text Metrics 0.1.0
 
 * Initial release.